+++ app/src/main/AndroidManifest.xml
@@ -51,10 +51,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".testCache.TestCacheActivity"
-            android:theme="@style/AppTheme.DefaultTheme"
-            android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
-        <activity
             android:name=".testSQL.TestSQLActivity"
             android:theme="@style/AppTheme.DefaultTheme"
             android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
