+++ base/build.gradle
@@ -26,6 +26,8 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
+    implementation project(':utils')
+
 //    implementation "com.dvsnier:base:$rootProject.base_version"
     api("com.dvsnier:common:$rootProject.common_version") {
         exclude group: 'com.android.support', module: 'appcompat-v7'
@@ -44,9 +46,10 @@ dependencies {
     implementation "com.android.support:support-annotations:$rootProject.support_annotations_version"
     //noinspection GradleDependency
     implementation "com.jakewharton:butterknife:$rootProject.butterknife_version"
+    implementation "com.jakewharton:disklrucache:$rootProject.disklrucache_version"
     annotationProcessor "com.jakewharton:butterknife-compiler:$rootProject.butterknife_compiler_version"
 
-    testImplementation "junit:junit:$rootProject.junit_version"
+    implementation "junit:junit:$rootProject.junit_version"
 
     androidTestImplementation "com.android.support.test:runner:$rootProject.runner_version"
     androidTestImplementation "com.android.support.test.espresso:espresso-core:$rootProject.espresso_core_version"
