+++ base/src/main/AndroidManifest.xml
@@ -1,12 +1,15 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.dvsnier.base.flavor">
 
-
     <application
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
             android:name=".crash.TestCrashHandleActivity"
+            android:theme="@style/AppTheme.DefaultTheme"
+            android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
+        <activity
+            android:name=".cache.TestCacheActivity"
             android:theme="@style/AppTheme.DefaultTheme"
             android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
     </application>
